% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CuaGL.R
\name{CGRL}
\alias{CGRL}
\title{Cuadrado Grecolatino}
\usage{
CGRL(respuesta, tratamiento, fila, columna, l_griega, data)
}
\arguments{
\item{respuesta}{(string) nombre de la variable respuesta}

\item{tratamiento}{nombre de la variable tratamiento}

\item{fila}{nombre de la variable fila}

\item{columna}{nombre de la variable columna}

\item{l_griega}{nombre de la variable l_griega}

\item{data}{(\code{data.frame}) Tabla de datos en formato largo con los datos
del tratamiento, filas, columnas, letras griegas y respuesta.}
}
\value{
Devuelve una tabla en formato \code{data.frame} con los cálculos
 correspondientes al análisis de varianza.
}
\description{
Realiza un cuadrado grecolatino
}
